+++ conformance/utils.go
@@ -1,0 +1,32 @@
+package main
+
+import (
+	"encoding/json"
+	"fmt"
+)
+
+func convert(i interface{}) interface{} {
+	switch x := i.(type) {
+	case map[interface{}]interface{}:
+		m2 := map[string]interface{}{}
+		for k, v := range x {
+			m2[k.(string)] = convert(v)
+		}
+		return m2
+	case []interface{}:
+		for i, v := range x {
+			x[i] = convert(v)
+		}
+	}
+	return i
+}
+
+// printJSON pretty prints a struct as JSON
+func printJSON(i interface{}) {
+	i = convert(i)
+	see, err := json.MarshalIndent(i, "", "   ")
+	if err != nil {
+		fmt.Printf("error printing JSON: %v\n", err)
+	}
+	fmt.Println(string(see))
+}
